Update @turbowarp/types (#513)
Development server users will need to run `npm ci` after pulling this.
types-tw is now installed as a git repository instead of an npm package: this reduces friction for updates considerably
Many new types are added for the new APIs that now exist. Most of the weird things in lab/text are now fully type checked.