Box2D: Make boxed and floor stage adapt to actual stage size
This commit makes the boundaries created when the stage type is set adapt to the stage size rather than use constants suitable for a 480x360 stage. It also adds comments to the creation of the body boxes, since it took me time interpreting the constants and instructions there, so I think the comments serve a useful purpose.