Box2D: Correct ceiling constants
When I calculated this constant I used `1000 - 360` because the ceiling was originally at `1000` and the height of the stage is `360` by default. However, I use the top stage bound and not the stage height when calculating the position for the ceiling. In this commit I recalculated the constant with `1000 - 180` because 180 is where the stage top bound is by default.