Lily/ClonesPlus: revert breaking change (#2110)
https://github.com/TurboWarp/extensions/commit/5825ea724fe303963adb8aa4af75ce75094bd38f
This change breaks projects made with previous versions of the
extension. This is was a change in functionality!

One of my projects is affected by this change, because it is built with
clones starting with empty local variables in mind. This should be
handled differently. I dont know how this was approved. You cant change
stuff like this after such a long time of them working this way, without
making sure that old projects are unaffected