Lily/Skins: Fix some fatal loading bugs. (#2128)
This causes a lot of errors if it happens repeatedly and can lead to
project corruption if handled incorrectly.
This fixes some of thoes loading bugs. fixes: #2127

The issue arises from the fact that you can rename / delete / cancel the
loading of a costume / skin while its still loading, and the renderer
throws and error if that happens at a bad time, Skins doesnt handle
this, this aims to fix that issue

---------

Co-authored-by: DangoCat[bot] <dangocat@users.noreply.github.com>