cloudlink: make every (non-localhost) server ID behave like 7 (#2164)
this is a bit of a "PR made slightly out of anger".

"use server 7" is almost memetic in the TW server, and some people think
CL is broken because people try to connect to e.g server ID 1 and it
doesn't work. 7 is the only server left anyways.

(0 still remains as localhost though)