Extension gallery enhancement: add search bar (#2055)
This is a small pull request that adds a client-side JS search bar to
the extension gallery. It searches by the contents of the titles and
descriptions. It doesn't require that a result that starts with the
search query, only that the results' strings contain the search query.

Potentially could also add sort/filter too


https://github.com/user-attachments/assets/69752cae-142b-4d35-a489-3467fe1c8a71