Validate import URLs are immutable (#2326)
Need to pin imports to an exact commit (git repository) or release (npm
package) so that they don't change unexpectedly. Hashes are checked at
build time so that would break the build. Pulling in updates without
vetting has risks related to security and unexpected breaking changes so
we should do it with some intention.